requirements.txt
amqp==5.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
idna==3.7
kombu==5.3.7
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
six==1.16.0
telebot==0.0.5
tradingview-ta==3.3.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
